ua_restore.cc: Remove unnecessary check for full backups
The check `rx->JobTDate == 0` is only true if the temporary table temp
is empty. Since we only reach this point if a full backup was found
and added to the temp table, this check will always be true since
adding a differential backup to the table cannot remove a full backup.