ua_restore.cc: Issue hint to restore from archives
When try restore from normal backup jobs where no suitable job was
found but a suitable archive job exists, give the user a hint to try
to restore from an archive instead (and vice versa).

This hint is currently not displayed when bareos tries to find a job
that contains a given amount of files and fails.