ua_restore: Fix bad tests
To test the behaviour of restore when no suitable backup is found but
a suitable archive does exist, we merge the simple backup & restore
tests. simple-backup-and-restore is always run first so we can
reliably say that only one backup job exists.  This way we can easily
control wether a backup job can be found or not.