systemtests: Create restore-test directory
To test some restore behaviour it is important for the tests to know
what backup jobs exist exactly.  To ensure that this is always the
case this new directory will include only one test that creates
backups, called create-backup, and every other test can then work with
only those backups.

One example of one such test is simple-archive-restore.  In this test
we also want to check wether a hint is displayed to the user to try
restoring from archive files if they are trying to run a restore but
no normal backup job exists while a appropriate archive job exists.
For this to work correctly the test has to be able to assume that it
knows for certain that 1) an archive job exists and 2) no other backup
job exists.