bfile: use kInvalidFiledescriptor
We should not assign a valid file descriptor to a bfd that was not
initialized yet.
Because of this IsBopen(bfd) curretly returns true (on unix systems)
if bfd was a newly created fd on which binit() was not called.
On windows systems it already correctly returns false in that
situation.  Now it always returns false.

Also uses kInvalidFiledescriptor when checking for invalid fds to stay
consistent.