bfile: add SEQUENTIAL and NOREUSE fadvise
We add the hints to each RDONLY/WRONLY opened file since we
just want to read/write the file once sequentially and send it to the
sd.

The sequential hint doubles the amount of readahead that the os will do
for us.  The noreuse hint does not do anything at the moment (on
linux), but might come in handy in the future.

bfile: add NOREUSE advice