find: rename SetupFlags -> SetupLastOptionBlock
Since this is basically what it does.  Some items are actually merged
between the different blocks.

Also added a call inside SendFiles since StripPath (which is used in
sending) is only set here and it needs to be correct before sending.
A better idea would be to set it in AcceptFile instead, but sadly the
call to AcceptFile happens _after_ the accurate check which uses the
StripPath bad StripPath value :(.
This still does not faithfully recreate the strippath value since now
it always takes it from the last option block whereas before it took
it from the last _used_ option block.