backup: read data serially sometimes
We cannot be sure that bread is "threadsafe" when its not done in
core; for example you cannot call python functions from other threads
than the main thread.  In situations like these, and when the file is
very small we need to ensure that the reading is done serially in the
main thread.

Technically we could move the reading part into the main thread and
the sending part into a worker thread which would solve the python
issue;  This still needs to be checked.