backup: allow non-network-buffer-size readsize
Currently the size we use for the network buffer is the same size as
the size we use for the read buffer.  This change divorces those two
concepts.  At the moment we still require that network buffer size
evenly divides read buffer size.  It should not be hard to remove that
requirement if we allow ourselves some copying.