compat: add cache invalidation
We need to invalidate the conversion cache after creating the vss
snapshots, since vss changes how paths are translated/converted and
the cache sometimes returns old values because of that.

Currently only the main thread is active in this stage so its enough
to only invalidate its cache.  As soon as other threads start being
active before vss is created we also need to invalidate their caches.