compat: remove com_security_mutex
We now just use a static object.  The initialization is guaranteed to
be thread safe and will happen on the first call to
InitializeComSecurity().  This slightly changes behaviour: we now do
not retry whenever the previous call failed; since the only way for
InitializeComSecurity() to fail is for the system to run out of
memory (according to the documenentation), this should not be a
concern.

(cherry picked from commit d2e90e8e09eee1c18d3137ca6721f6ae6a00cb8b)