unit-tests: rename vss conversion functions; fix strncat/cpy
Since strncat can write up to n+1 bytes (n bytes from src + null
terminator), we need to ensure that dest can hold up to that many
bytes!
Another issue is that strcpy may not leave the buffer with a zero
terminator.  To ensure that this is always the case now we just use
strncat twice!