consolidate.cc: keep keep original job data
when getting the client of the different jobs that need to be
consolidated, the `clientid` of the `job` record of the original
consolidation job gets overridden by that of the jobs being
consolidated. At the end of the operation, the name of the client of
the last job being treated persists as the original Consolidate job
client.

Other data also gets overridden, so the current fix saves the resource
and job record members of the OG Consolidation jcr before starting
the loop and brings them back during cleanup

(cherry picked from commit e43e9ba526f4a714b3cc79ebe2ed31ad73fa4c4d)