time-stamps: replace std::thread::id
std::thread::id is only guaranteed to be unique among currently
running threads.  While this is fine for now it will cause issues when
jobs create and delete threads on demand since they may have the same
ids.  To prevent this we now use an atomic counter and a thread local
variable for each thread that is actually unique as even if a thread
id is reused the threads will still have different thread locals!