time_stamps: add simple time stamp keeping
The idea is that you have a global TimeKeeper, which keeps a list of
all events that have happened (through thread local ThreadTimeKeepers),
that accepts a visitor-like ReportGenerator, which will be fed every
event that TimeKeeper has access too.

This can then be used to build any number of needed reports.