dedup-dev: fix d_open with mode & O_CREAT
O_CREAT only means "create if does not exist" and does not fail if the
file already exists.  As such (since its needed by the relabel
command) we try to create the volume if O_CREAT is set but if we fail,
we just fallthrough to the normal "open r/w" case.