dedup-dev: change code structure a lot
Instead of blockfiles, recordfiles, and datafiles getting a path and
an open mode, they instead get an already openend file descriptor.
This removes a lot of duplicate code and makes it possible to have
"true" constructors instead of the split constructor + open_inside
pair.

This is accomplished by splitting up the volume_config into two parts:
One part deals with the layout of the volume, that is which files it
includes and what is on those files, and one part that actually
contains the in memory representation of that layout (i.e. the opened files).