filed: allow plugins to (re-)set since-time
This changeset enables a few things that were not possible before:

override save_time in save_pkt, allowing a plugin to reset the effective
since-time on a file-by-file basis.

using SetValue() in a python plugin you can now reset since-time, which
will also set or clear the incremental flag as needed.
This allows a python plugin to call `SetValue(bVarSinceTime, 0)` to
make sure the core will not decide to skip any of the files it provides.