thread-pool: split thread_pool into thread_pool+work_group
Previously thread_pool took care of two different concerns:
1) Being able to reuse threads for work
2) Sharing threads for concurrent non-blocking computations

To be able to support both of these concerns we needed to add a lot of
complexity to the class.  Splitting it into two, a thread_pool, that
only lets us borrow threads for a while, and a work group, with which
you can share threads to do non-blocking computations, made the code
much easier to understand.