python-fd: provide config-file handling
this patch extends BareosFdPluginBaseclass to provide a standard way to
work with plugin configuration files.

When the plugin options parser encounters one of the options
defaults_file or overrides_file, it will treat the values as a path to
an ini-style configuration file and will try to read it. The resulting
options are added with precedence overrides_file, plugin definition,
defaults_file.
The code will consume the defaults_file and overrides_file options, so
the plugin will not see them later.