python-fd: avoid colons in encoded values
We now encode with b85encode instead of a85encode. The major difference
is the set of characters used to represent the result. While a85encode
will use a colon (:) that is also used as an option separator in the
plugindef, b85encode will not.
As a result values encoded with b85encode will be simpler to use.