tree-save: fix hardlink detection
We have to save the information on whether a file is the "original"
file or if its just a "link".  To do that we add the offset to the
original to every hardlinked file so that we can detect the actual
original by checking whether the node points to itself.