filed: fix not setting up thread specific jcr
When using client initiated connection we do not call the normal
procedure but instead directly call process_director_commands.  This
caused us to not correctly initialise the threadspecific jcr, which is
used in 1) debug messages and 2) for vss name lookup

This caused us to print bad jobids in debug messages as well as not
using the vss system at all.