resume: exclude "visible" files
We take advantage of the fact that the filedaemon only sends
directories after all their children was sent.  So if a directory is
contained in the directory tree of the job, we can just tell the
filedaemon to skip that directory without recursing into it.
Obviously the same is true for files as well.  As such we only need to
tell the filedaemon about very few entries for resuming since we only
need to send the uppermost directories and not their children.