compression: fix use of thread_local
Our win32-cross toolchain does not handle thread_local correctly.
This can load to crashes because it may free the storage before
destructing the object inside the storage.

See here for reference: https://sourceforge.net/p/mingw-w64/bugs/727/

This is extremely easy to reproduce if run in a debugger (gdb in my
case).  In the wild this does not happen 100%.

As such thread_local was replaced by a manually managed pthread_key.