read-records: fix not respecting bsr order
If we have a bsr chain

Vol 1 Part 1 -> Vol 2 -> Vol 1 Part 2

We have to absolutely make sure we follow this when reading, i.e. only
read the first part of vol 1, then read vol 2 and only then read the
second part of vol 1; otherwise this will lead to very bad restores!

The current hack is the following: We list all the volumes in the
correct order, with duplicates.  And during the match bsr calls we
make sure that we only look at the bsrs that are contigiously on that
volume, i.e. in the example above we would first only look at bsr 1,
then only at bsr 2 and only then we would look at bsr 3.