dird: keep track of runtime status via name
without this patch a new configuration resource would also create a new
runtime status. This lead to all sorts of issues. Including, but not
limited to, wrong bookkeeping of concurrent jobs.
Instead of creating a new runtime status, we now pick an existing one
for the resource with the same name and only create a new one if we do
not have a status yet.

(cherry picked from commit 0275871fadc82c7071320e0df9d93af2d0c264ca)