grpc: fix blocking behaviour of socket
We had a deadlock as we only send the number of bytes written after
everything was written, but we could not write everything to the
socket as the buffer is pretty small by default and the reader does
not start emptying it until we told him how many bytes he should
expect.

This is fixed by sending the reader constant updates in a stream
answer about how many bytes i actually wrote to the socket.