ua-output: fix not printing what bareos does
Its important that error messages output what bareos thinks and not
what the user inputed (as the user knows that himself).

This caused a weird issue on windows when 'show job=<jobname>' was
interpreted as 'show jobdefs=<jobname>' which returned the error

job resource <jobname> not found.

Which is extremely misleading, as bareos never searched for a job in
that case.  If the error message would have been instead

jobdefs resource <jobname> not found.

then the user at least notices that bareos did something wrong and try
some other input instead of trying to understand why the job was not
found.