compat: fix handling of relative paths
Our fix for invalid paths messed up our handling for relative paths as
'.' was treated as an invalid character and was replaced.  This caused
the '.' that glob() generated to end up in an absolute path, which was
is not valid.

We can now also reenable the parser test as this was the only reason
that it did not work.