grpc: fix crash
nullptr RestoreObjects are sent to every plugin, which will grpc to
try to pass them to its child, even if it does not exist.

Now these events are simply cached, so that the child can take care of
them when it gets initialized.

We should probably just drop these, as they dont really matter and
might just confuse the plugin...