vbackup: add check to ensure a file was actually selected
Previously we could leave the GetFileList() call without adding any
files to the bsr.  This would cause the bsr jobid to stay at 0 and
AddVolumeInformationToBsr would then print a very cryptic error
message (as it didnt know it should look for a specific jobid, but
instead search for the (unset) job name).

This fixes this by making sure we exit early if we detect that no file
was inserted into the bsr.